[proj2] Remove transmission delay in initialize_neighbor test
The test links two routers and checks if they share routing tables
immediately. However, when there are many routes to share, this can take
a while because transmission delay is 0.1 seconds by default, causing
the test to fail if the route to h2 is not the very first one shared.

This commit solves the problem by setting transmission delay to 0 for
the new link.