[proj4] Updated tests to work on Windows.
Previously, the tests failed on some Windows machines due to
Python's automatic conversion of newlines.  This caused unix
newlines ("\n") to be converted to Windows newlines ("\r\n")
when writing out data that had passed through the network, which
in turn would cause our comparision of files at the end of the
test to fail.  This commit adds the binary option to all file
opening, which means files are read as binary (rather than text)
data, and no newline conversion will occur.